+++ arch/arm/src/lpc31xx/lpc31_timerisr.c
@@ -111,7 +111,7 @@ void up_timer_initialize(void)
       freq   >>= 4;
     }
 
-  load = ((freq * (uint64_t)10000) / 1000000);
+  load = ((freq * 10000) / 1000000);
   putreg32((uint32_t)load, LPC31_TIMER0_LOAD);
 
   /* Set periodic mode */
