+++ arch/risc-v/src/esp32c3-legacy/esp32c3_tickless.c
@@ -453,8 +453,8 @@ int IRAM_ATTR up_timer_start(const struct timespec *ts)
       up_timer_cancel(NULL);
     }
 
-  cpu_ticks = SEC_2_CTICK((uint64_t)ts->tv_sec) +
-              NSEC_2_CTICK((uint64_t)ts->tv_nsec);
+  cpu_ticks = SEC_2_CTICK(ts->tv_sec) +
+              NSEC_2_CTICK(ts->tv_nsec);
 
   up_tmr_setcounter(cpu_ticks);
   g_timer_started = true;
