+++ arch/x86_64/src/intel64/intel64_perf.c
@@ -64,7 +64,7 @@ void up_perf_convert(clock_t elapsed, struct timespec *ts)
 
   ts->tv_sec  = elapsed / g_x86_64_timer_freq;
   left        = elapsed - ts->tv_sec * g_x86_64_timer_freq;
-  ts->tv_nsec = NSEC_PER_SEC * (uint64_t)left / g_x86_64_timer_freq;
+  ts->tv_nsec = NSEC_PER_SEC * left / g_x86_64_timer_freq;
 }
 #endif
