+++ arch/xtensa/src/esp32/esp32_wifi_adapter.c
@@ -2263,7 +2263,7 @@ static void wifi_rtc_disable_iso(void)
 
 int64_t esp32_timer_get_time(void)
 {
-  return (int64_t)esp_hr_timer_time_us();
+  return esp_hr_timer_time_us();
 }
 
 /****************************************************************************
